@@ -20,14 +20,10 @@
     "test:e2e": "jest --config ./test/jest-e2e.json"
   },
   "dependencies": {
-    "@nestjs/common": "^9.0.0",
-    "@nestjs/core": "^9.0.0",
-    "@nestjs/platform-express": "^9.0.0",
-    "@nestjs/serve-static": "^3.0.0",
     "@nestjs/typeorm": "^10.0.0",
     "class-validator": "^0.14.0",
     "dotenv": "^16.3.1",
-    "reflect-metadata": "^0.1.13",
+    "reflect-metadata": "^0.2.2",
     "rxjs": "^7.2.0",
     "sqlite3": "^5.1.6",
     "typeorm": "^0.3.17",
@@ -35,27 +31,31 @@
   },
   "devDependencies": {
     "@fortawesome/fontawesome-free": "^6.2.1",
-    "@nestjs/cli": "^9.0.0",
-    "@nestjs/schematics": "^9.0.0",
-    "@nestjs/testing": "^9.0.0",
+    "@nestjs/cli": "^10.3.2",
+    "@nestjs/common": "^10.3.7",
+    "@nestjs/core": "^10.3.7",
+    "@nestjs/platform-express": "^10.3.7",
+    "@nestjs/schematics": "^10.1.1",
+    "@nestjs/serve-static": "^4.0.2",
+    "@nestjs/testing": "^10.3.7",
     "@types/express": "^4.17.13",
-    "@types/jest": "29.2.4",
-    "@types/node": "18.11.18",
-    "@types/supertest": "^2.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "eslint": "^8.0.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "jest": "29.3.1",
-    "prettier": "^2.3.2",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.12.6",
+    "@types/supertest": "^6.0.2",
+    "@typescript-eslint/eslint-plugin": "^7.6.0",
+    "@typescript-eslint/parser": "^7.6.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "jest": "^29.7.0",
+    "prettier": "^3.2.5",
     "source-map-support": "^0.5.20",
     "supertest": "^6.1.3",
-    "ts-jest": "29.0.3",
+    "ts-jest": "^29.1.2",
     "ts-loader": "^9.2.3",
     "ts-node": "^10.0.0",
-    "tsconfig-paths": "4.1.1",
-    "typescript": "^4.7.4"
+    "tsconfig-paths": "^4.2.0",
+    "typescript": "^5.4.4"
   },
   "jest": {
     "moduleFileExtensions": [
