@@ -1,31 +1,31 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import Controllers from "./controllers/index"
-import Services from "./services/index"
-import Entities from "./entities/index"
+import Controllers from './controllers/index';
+import Services from './services/index';
+import Entities from './entities/index';
 
 // ----------------------------------- Vue.js -----------------------
-import { ServeStaticModule} from '@nestjs/serve-static'; //For Vue.js
+import { ServeStaticModule } from '@nestjs/serve-static'; //For Vue.js
 import { join } from 'path'; //For Vue.js
 
 @Module({
   imports: [
-    ServeStaticModule.forRoot({ 
+    ServeStaticModule.forRoot({
       rootPath: join(__dirname, '../..', 'front-end/dist'), //For Vue.js
     }),
 
     TypeOrmModule.forRoot({
       type: 'sqlite',
       database: './database/db.sqlite', // Path of SQLite file
-      entities: Entities, /* List of Entity for TypeORM*/
+      entities: Entities /* List of Entity for TypeORM*/,
       synchronize: true, // Attenzione: Remeber to put it false when after the production! //TODO mettilo false o perdi tabelle
+      autoLoadEntities: true,
+      keepConnectionAlive: true, // Attenzione: Remeber to put it false when after the production!
     }),
     TypeOrmModule.forFeature(Entities),
   ],
   controllers: Controllers,
   providers: Services,
 })
-
-
 export class AppModule {}
