@@ -1,20 +1,17 @@
-
 import { AppController } from '../app.controller';
 import { CategoryController } from '../database/category/category.controller';
-import { CostController } from '../database/cost/cost.controller';
-import { TypeController } from '../database/type/type.controller';
 import { WalletController } from '../database/wallet/wallet.controller';
-import { IncomeController } from '../database/income/income.controller';
 import { AmortizationController } from '../database/amortization/amortization.controller';
+import { TransactionController } from '../database/transaction/transaction.controller';
+import { MovementController } from 'src/database/movement/movement.controller';
 
 const Controllers = [
-    AppController,
-    CostController,
-    CategoryController,
-    TypeController,
-    WalletController,
-    IncomeController,
-    AmortizationController,
+  AppController,
+  CategoryController,
+  WalletController,
+  AmortizationController,
+  TransactionController,
+  MovementController,
 ];
 
-export default Controllers
+export default Controllers;
