@@ -1,37 +1,66 @@
-import { Controller, Get, Post, Delete, Body, Param, Query } from '@nestjs/common';
-import { AmortizationService } from './amortization.service';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Param,
+  Patch,
+  Delete,
+} from '@nestjs/common';
 import { Amortization } from './amortization.entity';
+import { AmortizationService } from './amortization.service';
 
-@Controller('amortizations')
+@Controller('amortization')
 export class AmortizationController {
   constructor(private readonly amortizationService: AmortizationService) {}
 
   @Get()
-  async getAmortization(@Query('startDate') startDate: string, @Query('endDate') endDate: string): Promise<Amortization[]> {
-    if (startDate && endDate) {
-      const parsedStartDate = new Date(startDate);
-      const parsedEndDate = new Date(endDate);
-
-      return this.amortizationService.findAllAmortizationByDate(parsedStartDate, parsedEndDate);
-    } 
-    else {
-      return this.amortizationService.findAllAmortization();
+  async findAll(): Promise<Amortization[]> {
+    try {
+      return await this.amortizationService.findAll();
+    } catch (error) {
+      throw error;
     }
   }
 
-  @Get(':id') // Aggiunta della route per ottenere una riga tramite ID
-  async getAmortizationById(@Param('id') id: number): Promise<Amortization> {
-    return this.amortizationService.findAmortizationById(id);
+  @Get(':id')
+  async findOne(@Param('id') id: string): Promise<Amortization> {
+    try {
+      return await this.amortizationService.findOne(+id);
+    } catch (error) {
+      throw error;
+    }
   }
 
-  @Post(':id')
-  async updateAmortization(@Param('id') id: number, @Body() amortization: Amortization): Promise<Amortization> {
-    return this.amortizationService.updateAmortization(id, amortization);
-  }
-  
-  @Delete(':id') // Aggiunta della route per eliminare una riga tramite ID
-  async deleteAmortizationById(@Param('id') id: number): Promise<void> {
-    await this.amortizationService.deleteAmortization(id);
+  @Post()
+  async create(
+    @Body() amortizationData: Partial<Amortization>,
+  ): Promise<Amortization> {
+    try {
+      return await this.amortizationService.create(amortizationData);
+    } catch (error) {
+      throw error;
+    }
   }
 
+  @Patch(':id')
+  async update(
+    @Param('id') id: string,
+    @Body() amortizationData: Partial<Amortization>,
+  ): Promise<Amortization> {
+    try {
+      return await this.amortizationService.update(+id, amortizationData);
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  @Delete(':id')
+  async delete(@Param('id') id: string): Promise<void> {
+    try {
+      await this.amortizationService.delete(+id);
+    } catch (error) {
+      throw error;
+    }
+  }
 }
