@@ -1,99 +1,135 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { Like, Repository } from 'typeorm';
 import { Wallet } from './wallet.entity';
+import { logger } from '../../module/logger';
 
 @Injectable()
 export class WalletService {
   constructor(
     @InjectRepository(Wallet)
-    //private readonly costRepository: Repository<Cost>,
-    private walletRepository: Repository<Wallet>,
+    private readonly walletRepository: Repository<Wallet>,
   ) {}
 
   async findAllWallet(): Promise<Wallet[]> {
-    return this.walletRepository.find();
+    try {
+      return await this.walletRepository.find();
+    } catch (error) {
+      logger.error(`Error while fetching all wallets: ${error}`);
+      throw new BadRequestException('Error while fetching all wallets');
+    }
   }
 
   async findWalletById(id: number): Promise<Wallet> {
-    return this.walletRepository.findOne({ where: { id } });
+    try {
+      return this.walletRepository.findOne({ where: { id } });
+    } catch (error) {
+      logger.error(`Error while fetching all wallets: ${error}`);
+      throw new BadRequestException(
+        `Error while fetching wallet with ID: ${id}`,
+      );
+    }
   }
 
   async createWallet(wallet: Wallet): Promise<Wallet> {
     try {
-        await this.validateUniqueness(wallet);
-        await this.validateInput(wallet);
-        return await this.walletRepository.save(wallet); 
+      await this.validateInput(wallet);
+
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      const { id, ...newWallet } = wallet;
+      return await this.walletRepository.save(newWallet);
     } catch (error) {
       if (error instanceof BadRequestException) {
         throw error; // L'eccezione di univocità è già stata gestita, rilanciala
       }
-      throw new BadRequestException('Errore durante la creazione del wallet.');
+      logger.error(`Error while creating a new wallet: ${error}`);
+      throw new BadRequestException('Error while creating a new wallet');
     }
   }
 
   async updateWallet(id: number, wallet: Wallet): Promise<Wallet> {
     try {
-      await this.validateInput(wallet);
+      await this.validateInput(wallet, id);
+
       const updatedCategory = await this.walletRepository.update(id, wallet);
 
       // Verifica se l'aggiornamento ha avuto successo
       if (updatedCategory.affected === 0) {
-        throw new NotFoundException(`Wallet con ID ${id} non trovata.`);
+        logger.error(`Error wallet don't found ID: ${id}`);
+        throw new NotFoundException(`Wallet con ID ${id} non trovato.`);
       }
       return wallet;
-
     } catch (error) {
       if (error instanceof BadRequestException) {
         throw error; // L'eccezione di univocità è già stata gestita, rilanciala
       }
-      throw new BadRequestException('Errore durante la modifica del wallet.');
+      logger.error(`Error while updating a new wallet: ${error}`);
+      throw new BadRequestException('Error while updating a wallet');
     }
   }
-  
 
-  //GESTIRE BADREQUESTEXCEPTION se voglio eliminare tutto il wallet con i costi o no
+  //TODO GESTIRE BADREQUESTEXCEPTION se voglio eliminare tutto il wallet con i costi o no
   async deleteWallet(id: number): Promise<void> {
     try {
       await this.walletRepository.delete(id);
     } catch (error) {
-  
-      if (error instanceof Error && error.message.includes('FOREIGN KEY constraint failed')) {
-        throw new BadRequestException('Impossibile eliminare il wallet: è ancora referenziata a qualche cost o income, elimina quelli prima.');
+      if (
+        error instanceof Error &&
+        error.message.includes('FOREIGN KEY constraint failed')
+      ) {
+        throw new BadRequestException(
+          'Impossibile eliminare il wallet: è ancora referenziata a qualche cost o income, elimina quelli prima.',
+        );
       }
-  
+
       if (error instanceof BadRequestException) {
         throw error; // L'eccezione di univocità è già stata gestita, rilanciala
       }
-  
-      throw new BadRequestException('Errore durante la eliminazione del wallet.');
+
+      logger.error(`Error while deleting a wallet: ${error}`);
+      throw new BadRequestException('Error while deleting a wallet');
     }
   }
-  
-  
 
-  async validateUniqueness(wallet: Wallet): Promise<void> {
-    const description = wallet.description;
-    const existingCategory = await this.walletRepository.findOne({
-      where: { description },
+  async validateUniqueness(wallet: Wallet, id = null): Promise<void> {
+    const name = wallet.name;
+    const existingWallet = await this.walletRepository.findOne({
+      where: [{ name }, { name: Like(`%${name}%`) }],
     });
-  
-    if (existingCategory) {
-      throw new BadRequestException('Errore, la descrizione deve essere univoca.');
+
+    if (existingWallet && id != existingWallet.id) {
+      throw new BadRequestException('The name must be unique');
     }
   }
 
-  async validateInput(wallet: Wallet): Promise<void> {
-    if (!wallet.description || wallet.description.trim() === '') {
-      throw new BadRequestException('Errore, la descrizione non può essere nulla.');
-    }
-    
-    if (!wallet.money || wallet.money < 0) {
-      throw new BadRequestException('Errore, i soldi non possono essere negativi.');
+  async validateInput(wallet: Wallet, id = null): Promise<void> {
+    if (!wallet.name || wallet.name.trim() === '') {
+      throw new BadRequestException('The name cannot be invalid.');
     }
 
-    // Arrotonda il valore money a due decimali
-    wallet.money = parseFloat(wallet.money.toFixed(2));
+    // Controlla se wallet.currency è un numero decimale
+    if (
+      isNaN(parseFloat(wallet.currency.toString())) ||
+      !isFinite(wallet.currency)
+    ) {
+      throw new BadRequestException('Currency must be a decimal number.');
+    }
+
+    // Converte wallet.currency in un numero decimale
+    wallet.currency = parseFloat(wallet.currency.toString());
+
+    if (wallet.currency < 0) {
+      throw new BadRequestException('Currency cannot be negative.');
+    }
+
+    // Arrotonda il valore currency a due decimali
+    wallet.currency = parseFloat(wallet.currency.toFixed(2));
+
+    //Checking uniquiness
+    await this.validateUniqueness(wallet, id);
   }
-
 }
