@@ -1,17 +1,9 @@
-import { Cost } from '../database/cost/cost.entity';
+import { Transaction } from 'src/database/transaction/transaction.entity';
 import { Category } from '../database/category/category.entity';
-import { Income } from '../database/income/income.entity';
-import { Type } from '../database/type/type.entity';
 import { Wallet } from '../database/wallet/wallet.entity';
 import { Amortization } from 'src/database/amortization/amortization.entity';
+import { Movement } from 'src/database/movement/movement.entity';
 
-const Entities = [
-    Cost,
-    Category,
-    Income,
-    Type,
-    Wallet,
-    Amortization,
-];
+const Entities = [Category, Wallet, Amortization, Transaction, Movement];
 
-export default Entities
+export default Entities;
