@@ -1,20 +1,17 @@
-
+import { TransactionService } from 'src/database/transaction/transaction.service';
 import { AppService } from '../app.service';
 import { CategoryService } from '../database/category/category.service';
-import { CostService } from '../database/cost/cost.service';
-import { TypeService } from '../database/type/type.service';
 import { WalletService } from '../database/wallet/wallet.service';
-import { IncomeService } from '../database/income/income.service';
 import { AmortizationService } from 'src/database/amortization/amortization.service';
+import { MovementService } from 'src/database/movement/movement.service';
 
 const Services = [
-    AppService,
-    CategoryService,
-    CostService,
-    TypeService,
-    WalletService,
-    IncomeService,
-    AmortizationService,
+  AppService,
+  CategoryService,
+  WalletService,
+  AmortizationService,
+  TransactionService,
+  MovementService,
 ];
 
-export default Services
+export default Services;
